fix(security): G112-G115/prevent HTTP slowloris and unsafe integer conversions
- configure HTTP read, write, header, and shutdown timeouts
- align unsigned database fields with uint64 domain types
- validate attachment dimension and duration boundaries
- prevent negative message sequence conversion
- add tests for numeric conversion boundaries

+++ internal/service/message_service.go
@@ -64,7 +64,7 @@ func (s *MessageService) SendMessage(
 		ParentID:       parentID,
 		Type:           model.MessageType(msgType),
 		Content:        &content,
-		Seq:            uint64(seqVal),
+		Seq:            seqVal,
 	}
 	if err := s.msgRepo.InsertMessage(ctx, arg); err != nil {
 		return nil, ae.Internal(err)
@@ -116,7 +116,7 @@ func (s *MessageService) SendMessageWithAttachment(
 		ParentID:       parentID,
 		Type:           model.MessageType(msgType),
 		Content:        &content,
-		Seq:            uint64(seqVal),
+		Seq:            seqVal,
 	}
 
 	if err := s.msgRepo.InsertMessage(ctx, arg); err != nil {
